Fix bla
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@36556 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/pet/petpia2.c
@@ -126,7 +126,7 @@ static uint8_t read_pa(void)
         log_message(mypia_log,
                     "read pia2 port A %x, parallel_bus=%x, gives %x.",
                     mypia.port_a, parallel_bus,
-                    ((parallel_bus & ~mypia.ddr_a)
+                    (unsigned int)((parallel_bus & ~mypia.ddr_a)
                      | (mypia.port_a & mypia.ddr_a)));
     }
